Tim ra cho hong tren card lien: mot mang so thuc 64 bit
Bo dam da chi dung ten:

    [ duoc  ] sim.fans()     (48, 12)   float64      <- phai la float32
    [ KHONG ] clamp(0, 8)
               RuntimeError: unknown error

Card lien khong lam duoc `clamp` tren so thuc 64 bit. Con 64 bit tu dau ra
thi la mot dong trong `turbo/sim.py`:

    self.scan_r = torch.full((R, N_LIDAR), P.LIDAR_MAX, device=device)

Khong ghi `dtype`. Tren CPU thi `torch.full` lay kieu mac dinh la 32 bit
nen khong ai thay gi suot tu dau den gio; tren card lien no lay kieu cua
chinh con so Python, ma so Python la 64 bit.

Va day la ly do no kho tim den vay: loi CHI hien ra o xe VUA DAT LAI. Quet
xong mot vong LiDAR la `_finish_scan` thay `scan_r` bang mang 32 bit va loi
bien mat. Nen:

  - muc 2 (chay 40 buoc, da quet xong may vong) -> qua
  - muc 4 (vua dat lai, moi quet 60/500 diem)   -> chet
  - `fans()` goi rieng                          -> "duoc", vi tra ve 64 bit
                                                   van tra ve duoc, chet la
                                                   o `clamp` sau do
  - doi cach gom tia, chay rieng, lo nho hon    -> deu vo ich, vi kieu so
                                                   khong doi

Sua: ghi ro `dtype=torch.float32` o moi cho tao mang co con so Python -
`scan_r`, hai cho trong `ops.fan_min`, va may cho trong bo do.

Chan khong cho quay lai, ba lop:

  - `ops.dtype_audit()`: soi moi tensor cua mot doi tuong, bao cai nao
    khong phai 32 bit / dung-sai / so nguyen.
  - `check.py` goi no ngay sau khi dung mo phong VA ngay sau khi dat lai xe
    - dung hai luc de lot nhat.
  - Hai bai kiem thu: mot bai soi trang thai ngay sau khi dat lai (chua quet
    vong nao), mot bai chay ca mot vong danh gia duoi `TorchFunctionMode` va
    bao loi neu BAT KY phep nao sinh ra so thuc 64 bit. Da thu dat lai dong
    cu vao: ca hai bai deu bat duoc.

Khong doi gi tren CPU - `torch.full` o do von da la 32 bit, nen moi so do
cu van nguyen gia tri.

+++ turbo/sim.py
@@ -131,7 +131,14 @@ class BatchSim:
         self.lox = torch.zeros(self.R, N_LIDAR, device=device)
         self.loy = torch.zeros(self.R, N_LIDAR, device=device)
         self.loth = torch.zeros(self.R, N_LIDAR, device=device)
-        self.scan_r = torch.full((self.R, N_LIDAR), P.LIDAR_MAX, device=device)
+        # `dtype` phai ghi RO. `torch.full` voi mot so Python tren card lien
+        # ra so thuc 64 BIT, va card do khong lam duoc `clamp` tren 64 bit.
+        # Cai buoc mo phong dau tien chet la vi day: chua quet xong vong nao
+        # thi `scan_r` van la mang goc nay, quet xong mot vong thi
+        # `_finish_scan` thay no bang mang 32 bit - nen loi chi hien ra o xe
+        # VUA DAT LAI, va bien mat sau mot vong quet.
+        self.scan_r = torch.full((self.R, N_LIDAR), P.LIDAR_MAX,
+                                 dtype=torch.float32, device=device)
         self.scan_b = torch.zeros(self.R, N_LIDAR, device=device)
         self.scan_ok = torch.zeros(self.R, N_LIDAR, dtype=torch.bool,
                                    device=device)
